Chore: Refactor server codes
