+++ server/routes/controllers/index.ts
@@ -1,0 +1,4 @@
+export interface propsType {
+  req: Request;
+  res: Response;
+}
